wrapper locally ready

+++ cs336_basics/train_wrapper.py
@@ -45,18 +45,19 @@ def train_model_TinyStories(d_model=512,h=16,d_ff=1344,vocab_size=10000,context_
     vocab_file = data_dir / "TinyStories_vocab.txt"
     merges_file = data_dir / "TinyStories_merges.txt"
 
-    print("start training BPE")
-    bpe_text=dataset["train"]["text"]
-    bpe_text_path = data_dir / "TinyStories_100k.txt"
-    with open(bpe_text_path, "w", encoding="utf-8") as f:
-        f.write("\n".join(bpe_text))
-    print("text saving successful")
-    tok = bpe_tokenizer.tokenizer_training(vocab_size, ["<|endoftext|>"])
-    vocab, merges = tok.train_tokenizer(str(bpe_text_path))
-    print("training BPE successful")
+    if not vocab_file.exists() and merges_file.exists():
+        print("Tokenizer not found — training BPE")
+        bpe_text=dataset["train"]["text"]
+        bpe_text_path = data_dir / "TinyStories_100k.txt"
+        with open(bpe_text_path, "w", encoding="utf-8") as f:
+            f.write("\n".join(bpe_text))
+        print("text saving successful")
+        tok = bpe_tokenizer.tokenizer_training(vocab_size, ["<|endoftext|>"])
+        vocab, merges = tok.train_tokenizer(str(bpe_text_path))
+        print("training BPE successful")
 
-    bpe_tokenizer.tokenizer.save_tokenizer(vocab, merges, str(vocab_file), str(merges_file))
-    print(f"saved tokenizer to {vocab_file} and {merges_file}")
+        bpe_tokenizer.tokenizer.save_tokenizer(vocab, merges, str(vocab_file), str(merges_file))
+        print(f"saved tokenizer to {vocab_file} and {merges_file}")
 
     print("start tokenizing")
     tokenizer=bpe_tokenizer.tokenizer.from_files(
